[WOR-1517] Remove workspace-level READ check  (#61)
Leo uses it's own SA identity to poll on resource operations within an MRG, such as welder and jupyter. However, READ is auth domain constrainable. When an auth domain is applied to the workspace, the SamResourceClient's  READ check will fail and Leo will not be able to perform the needed operations. Since there is no longer a direct policy applied to apps in a workspace, the inheritance from the workspace will be enough to ensure those removed from a workspace no longer have access to shared apps.